final-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "openai (>=2.21.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
